big update
reworked the ultrasonic sensor. so it doesn't misuse the systick timer. timer now gets used for timing the pulse, the trigger needs to be activated in software.

fixed an critical bug in the ledstrip driver, where the resetpulse initiation was selected a cylce to late, resulting in erratic behaviour.

fixed a padding issue in the unixfunciton.c

fixed several issues caused by improper typecasting

commented a function in the lcd driver, that caused to hang the system. UPDATE REQUIRED. TODO

switched the ledstrip to TPM2 instead of TPM0

moved interrupt handlers that share interrupts to a separate source and header file.
intterupt handlers that are used by one module can stay in that modules source files.

added a debug test function where the ledstrip will light up if an object is detected in a range of 10 cm.

added contintuous update function to the led strip.

moved common #includes and #defines to common.h

+++ tempusludicus/Core/Inc/unixFunction.h
@@ -11,7 +11,7 @@ typedef struct {
     uint16_t day;
     uint16_t hour;
     uint16_t minute;
-    uint8_t second;
+    uint16_t second;
 
 } datetime_t;
